# don't recheck events in `select_from_last_begin`