--- a/lib/fusuma.rb
+++ b/lib/fusuma.rb
@@ -56,7 +56,10 @@ module Fusuma
     end
 
     def initialize
-      @inputs = Plugin::Inputs::Input.plugins.map(&:new)
+      @inputs = Plugin::Inputs::Input.plugins.map do |cls|
+        cls.ancestors.include?(Singleton) ? cls.instance : cls.new
+      end
+      @inputs.push(Plugin::Inputs::TimerInput.instance)
       @filters = Plugin::Filters::Filter.plugins.map(&:new)
       @parsers = Plugin::Parsers::Parser.plugins.map(&:new)
       @buffers = Plugin::Buffers::Buffer.plugins.map(&:new)