--- a/lib/fusuma/plugin/detectors/hold_detector.rb
+++ b/lib/fusuma/plugin/detectors/hold_detector.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "./detector"
+require_relative "../inputs/timer_input"
 
 module Fusuma
   module Plugin
@@ -10,6 +11,7 @@ module Fusuma
         SOURCES = %w[gesture timer].freeze
         BUFFER_TYPE = "gesture"
         GESTURE_RECORD_TYPE = "hold"
+        Timer = Inputs::TimerInput.instance
 
         BASE_THERESHOLD = 0.7
 
@@ -21,16 +23,15 @@ module Fusuma
           hold_buffer = find_hold_buffer(buffers)
           return if hold_buffer.empty?
 
-          hold_events = hold_buffer.events
+          last_hold = hold_buffer.events.last
 
           timer_buffer = buffers.find { |b| b.type == "timer" }
-          timer_events = timer_buffer.events
+          last_timer = timer_buffer.events.last
 
           finger = hold_buffer.finger
-          holding_time = calc_holding_time(hold_events: hold_events, timer_events: timer_events)
+          holding_time = calc_holding_time(hold_events: hold_buffer.events, last_timer: last_timer)
 
-          @timeout ||= nil
-          status = case hold_events.last.record.status
+          status = case last_hold.record.status
           when "begin"
             if holding_time.zero?
               "begin"
@@ -42,16 +43,19 @@ module Fusuma
           when "end"
             "end"
           else
-            last_record = hold_events.last.record.status
+            last_record = last_hold.record.status
             raise "Unexpected Status:#{last_record.status} in #{last_record}"
           end
 
           repeat_index = create_repeat_index(finger: finger, status: status)
           oneshot_index = create_oneshot_index(finger: finger)
 
-          @timeout = nil if status == "begin"
-
-          if status == "timer"
+          if status == "begin" then
+            @timeout = nil
+            if threshold(index: oneshot_index) < Timer.interval then
+              Timer.wake_early(Time.now + threshold(index: oneshot_index))
+            end
+          elsif status == "timer"
             return if @timeout
 
             return unless enough?(index: oneshot_index, holding_time: holding_time)
@@ -100,9 +104,9 @@ module Fusuma
             .select_by_type(GESTURE_RECORD_TYPE)
         end
 
-        def calc_holding_time(hold_events:, timer_events:)
-          last_time = if !timer_events.empty? && (hold_events.last.time < timer_events.last.time)
-            timer_events.last.time
+        def calc_holding_time(hold_events:, last_timer:)
+          last_time = if last_timer && (hold_events.last.time < last_timer.time)
+            last_timer.time
           else
             hold_events.last.time
           end
@@ -110,7 +114,13 @@ module Fusuma
         end
 
         def enough?(index:, holding_time:)
-          holding_time > threshold(index: index)
+          diff = threshold(index: index) - holding_time
+          if diff < 0 then
+            true
+          elsif diff < Timer.interval
+            Timer.wake_early(Time.now + diff)
+            false
+          end
         end
 
         def threshold(index:)