--- a/lib/fusuma/plugin/detectors/pinch_detector.rb
+++ b/lib/fusuma/plugin/detectors/pinch_detector.rb
@@ -19,7 +19,7 @@ module Fusuma
         def detect(buffers)
           gesture_buffer = buffers.find { |b| b.type == BUFFER_TYPE }
             .select_from_last_begin
-            .select_by_events { |e| e.record.gesture == GESTURE_RECORD_TYPE }
+            .select_by_type(GESTURE_RECORD_TYPE)
 
           updating_events = gesture_buffer.updating_events
           return if updating_events.empty?