--- a/lib/fusuma/plugin/detectors/rotate_detector.rb
+++ b/lib/fusuma/plugin/detectors/rotate_detector.rb
@@ -24,16 +24,9 @@ module Fusuma
           updating_events = gesture_buffer.updating_events
           return if updating_events.empty?
 
-          oneshot_angle = if updating_events.size >= 10
-            updating_time = 100 * (updating_events[-1].time - updating_events[-10].time)
-            last_10 = gesture_buffer.class.new(updating_events[-10..-1])
-            last_10.sum_attrs(:rotate) / updating_time
-          else
-            updating_time = 100 * (updating_events.last.time - updating_events.first.time)
-            gesture_buffer.sum_attrs(:rotate) / updating_time
-          end
-
-          return if updating_events.empty?
+          updating_time = 100 * (updating_events[-1].time -
+                                 (updating_events[-10] || updating_events.first).time)
+          oneshot_angle = gesture_buffer.sum_last10_attrs(:rotate) / updating_time
 
           finger = gesture_buffer.finger