--- a/lib/fusuma/plugin/inputs/timer_input.rb
+++ b/lib/fusuma/plugin/inputs/timer_input.rb
@@ -1,20 +1,29 @@
 # frozen_string_literal: true
 
 require_relative "./input"
+require 'timeout'
 
 module Fusuma
   module Plugin
     module Inputs
       # libinput commands wrapper
       class TimerInput < Input
-        DEFAULT_INTERVAL = 0.3
+        include Singleton
+        DEFAULT_INTERVAL = 5
+        EPSILON_TIME = 0.02
         def config_param_types
           {
             interval: [Float]
           }
         end
 
-        attr_reader :pid
+        def initialize(*args, interval: nil)
+          super(*args)
+          @interval = interval || config_params(:interval) || DEFAULT_INTERVAL
+          @early_wake_queue = Queue.new
+        end
+
+        attr_reader :pid, :interval
 
         def io
           @io ||= begin
@@ -34,9 +43,17 @@ module Fusuma
 
         def timer_loop(writer)
           begin
+            delta_t = @interval
+            next_wake = Time.now + delta_t
             loop do
-              sleep interval
-              writer.puts "timer"
+              begin
+                Timeout.timeout(next_wake - Time.now) do
+                  next_wake = [@early_wake_queue.deq, next_wake].min
+                end
+              rescue Timeout::Error
+                writer.puts "timer"
+                next_wake = Time.now + delta_t
+              end
             end
           rescue Errno::EPIPE
             exit 0
@@ -45,14 +62,14 @@ module Fusuma
           end
         end
 
+        def wake_early(t)
+          @early_wake_queue.push(t + EPSILON_TIME)
+        end
+
         private
 
         def create_io
           IO.pipe
-        end
-
-        def interval
-          config_params(:interval) || DEFAULT_INTERVAL
         end
       end
     end