--- a/spec/fusuma/plugin/inputs/timer_input_spec.rb
+++ b/spec/fusuma/plugin/inputs/timer_input_spec.rb
@@ -10,26 +10,23 @@ module Fusuma
         before do
           @dummy_read = StringIO.new("dummy_read")
           @dummy_write = StringIO.new("dummy_write")
-          @input = TimerInput.new
+          @input = TimerInput.instance
           allow(@input).to receive(:create_io).and_return [@dummy_read, @dummy_write]
-          allow(@input).to receive(:fork)
-          allow(Process).to receive(:detach).with(anything)
+          allow(Thread).to receive(:new)
         end
 
         describe "#io" do
-          it { expect(@input.io).to eq @dummy_read }
-
           it "should call #create_io" do
             expect(@input).to receive(:create_io)
             expect(@input).to receive(:start)
-            @input.io
+            expect(@input.io).to eq @dummy_read
           end
         end
 
         describe "#start" do
           it {
-            expect(@input).to receive(:fork).and_yield do |block_context|
-              expect(block_context).to receive(:timer_loop).with(@dummy_read, @dummy_write)
+            expect(Thread).to receive(:new).and_yield do |block_context|
+              expect(block_context).to receive(:timer_loop).with(@dummy_write)
             end
             @input.start(@dummy_read, @dummy_write)
           }